# apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/client-service.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/messages": {
      "get": {
        "operationId": "MessageService_GetMessageStream",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/gomicrotestserviceGetListMessageResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of gomicrotestserviceGetListMessageResponse"
            }
          }
        },
        "tags": [
          "MessageService"
        ]
      },
      "post": {
        "operationId": "SaveMessage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gomicrotestserviceSaveMessageResponse"
            }
          },
          "default": {
            "description": "Error response",
            "schema": {
              "$ref": ".testservice.Error"
            }
          }
        },
        "tags": [
          "MessageService"
        ]
      }
    },
    "/api/v1/messages/{messageId}": {
      "get": {
        "operationId": "MessageById",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gomicrotestserviceSingleMessageResponse"
            }
          },
          "default": {
            "description": "Error response",
            "schema": {
              "$ref": ".testservice.Error"
            }
          }
        },
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MessageService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "gomicrotestserviceGetListMessageResponse": {
      "type": "object",
      "properties": {
        "messageId": {
          "type": "string"
        },
        "messageDescription": {
          "type": "string"
        }
      }
    },
...
    "gomicrotestserviceSaveMessageResponse": {
      "type": "object",
      "properties": {
        "messageId": {
          "type": "string"
        }
      }
    },
    "gomicrotestserviceSingleMessageResponse": {
      "type": "object",
      "properties": {
        "messageId": {
          "type": "string"
        },
        "messageDescription": {
          "type": "string"
        }
      }
    }
  }
}